fix: ensure task scores are strictly within (0, 1) range

+++ server/app.py
@@ -66,7 +66,7 @@ def step(req: StepRequest):
 
     return {
         "question": result.observation.question,
-        "reward": int(result.reward),
+        "reward": float(result.reward),
         "done": result.done
     }
